[UPDATE] implementation, tests, eqn

+++ test/test.js
@@ -191,7 +191,7 @@ describe( 'distributions-poisson-mgf', function tests() {
 			expected,
 			i;
 
-		data = new Float32Array( validationData.data );
+		data = new Float64Array( validationData.data );
 
 		expected = new Float64Array( validationData.expected.map( function( d ) {
 			return d === 'Inf' ? Infinity : d;
@@ -213,7 +213,7 @@ describe( 'distributions-poisson-mgf', function tests() {
 			'copy': false,
 			'lambda': validationData.lambda
 		});
-		expected = new Float32Array( validationData.expected.map( function( d ) {
+		expected = new Float64Array( validationData.expected.map( function( d ) {
 			return d === 'Inf' ? Infinity : d;
 		}) );
 		assert.strictEqual( actual, data );
